Reviewed PR #64, added a test for it and updated our testing scheme (tests were not working)
