+++ tests/data/vulnerable-package.json
